divide by a files
